Increase replicas and remove liveness probe
When the model gateway is on high load, all web processes are busy
serving external requests. The liveness probe is amongst those requests
which got timed out while waiting for a response. K8s monitor will pick
up these failures as the pod is unhealthy and trying to restart it. This
results in the pod being in a CrashLoopBackOff state.