Add buckets between 1 & 10
This is where most of our successful requests fall between.

According to the logs, the P95 for 2xx requests is about 2.5s, the p99
is about 3s.

For 4xx and other failures the p99 duration is about 0.5s.

Source:
https://log.gprd.gitlab.net/goto/945b79c0-fbc7-11ed-8afc-c9851e4645c0

With these extra buckets, we could maybe tighten up the satisfied
threshold for the apdex to 5s and have a tolerated threshold of
10s. While in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/5852/ we set a
tolerated threshold of 10s.