Add GCP quota exporter
This adds https://github.com/mintel/gcp-quota-exporter/ as exporter
for GCP quota. Sadly, this project is archived

However, it is the same image as we use for exporting these metrics in
our main cluster, and as a result, the metrics will be named the
same. So I think we should still use this one until we change these to
something new everywhere.

This requires a separate Google Service Account to be created that has
the right permissions to request the quota, and allowing the
Kubernetes service account assigned to the `gcp-quota-expoter` pod.

This is done by creating a new `gcp-quota-exporter` Google Service
account and assigning it a custom role with the following permissions:

- compute.projects.get
- compute.regions.list
- compute.instances.list
- serviceusage.quotas.get

Then we can assign the `workloadIdentityUser` role to the Kubernetes
service account for the pod.

Part of
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23861