Collect usage statistics over telemetry headers
This change extracts telemetry headers from /v2/completions requests and
stores them in telemetry log and Prometheus metrics. If the headers are
not present or the data is not valid, we will log the error and proceed
with serving the requests.

The telemetry log is output via a separate log stream named `telemetry`.