Deploy the runner in a separate environment
It is still dpeloyed in the `ai-assist-test` cluster, but in a
different environment.

This allows us to move the values for the runner, that require a
secret env variable to be present, into a separate file.

In turn, this allows us to diff other environments than the runner one
to still see changes.