Error on pytest `UserWarning`s
At present, some test cases may be skipped and a warning emitted if
some preconditions are not met. Combined with our use of the
`--disable-warnings` flag, this means the test suite can pass with
no indication to the user that something's wrong. Lest add a warning
filter that errors on `UserWarning` (which is the superclass of Pytest
warnings, see
https://happytest-apidoc.readthedocs.io/en/latest/_modules/_pytest/warning_types/)