feat: log editor language ID
Currently the Code Suggestions backend looks at the file extension to
determine what language the code might be. If the user overrides the
language setting in the client, the backend should know about it. This
commit only logs the `editor_lang` in the Prometheus language counters
so we can see how often the guessed language matches the editor's
language.

This requires an update to the client:
https://gitlab.com/gitlab-org/gitlab-vscode-extension/-/merge_requests/1004