Remove new line on generations with Anthropic
This removes new line prepending on model output for code generations
with Anthropic since we do not have a reliable way to detect when to
perform this processing. The false positive case, ie. we prepend new
line when shouldn't, is more severe than the false negative. User has to
perform more keystrokes to rectify the later.