Reduce Anthropic timeout
The existing/default timeout is 10 minutes which is really high. With
the new change in the client, timeout requests would be retried, which
would make the 10-minute timeout less usable.

This commit reduces timeout to 30 sec according to the current metrics
in production and allow us to take advantage of retry.