Expose configured concurrency limits for models
This allows us to configure limits for a certain engine and model by
setting the environment variable `MODEL_ENGINE_CONCURRENCY_LIMITS`
with a json value as follows:

```
{
  "vertex-ai": { "code-gecko@002": 7 },
  "anthropic": { "claude-2.0": 2, "claude-2.1": 7 }
}
```

When doing this, we will expose these limits in a Prometheus gauge
called `model_inferences_max_concurrent` soon as we've run an
inference for this model. When no value is set for a model, the gauge
is not exposed.

This allows us to combine this with the `model_inferences_in_flight`
gauge to see how much we're using at the time of sampling.

For https://gitlab.com/gitlab-com/runbooks/-/issues/143