Add documentation on inference limits
This documents what the inference limits are, where and on what
dashboards we can see them.

It also explains the different limits we have for Anthropic and
Vertex.

In the case of Anthropic, it also shows how these limits need to be
manually updated when models or limits change.

For
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/issues/391