feat: use multiprocess metrics automatically in Docker image
This is a follow-up merge request to
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/merge_requests/776.

If a user sets `WEB_CONCURRENCY` to a value > 1 so that multiple
worker processes are spawned, automatically enable Prometheus
multiprocess metrics by setting the `PROMETHEUS_MULTIPROC_DIR` to a
temporary directory. This will enable anyone running the Docker image
to scrape the metrics endpoint and get aggregate data from all the
workers.