chore: use CodeCompletionsLegacy for testing vertex
Previously we were using CodeGenerations for validating both Anthropic
and Vertex availability in a region. However this is not supported
using Vertex since !822

In this merge request, we iterate all of the predictions we get from a
model to build multiple `CodeSuggestionsOutput`s in a list, instead of
returning the first one from `PalmCodeGenBaseModel._generate`. This
breaks compatibility when consuming the model in
`CodeCompletions._handle_sync`.

This is not a configuration that is currently used by the AI-gateway
deployment, so we did not see this error in production. We should
decide what to do with that support in #493

In the meantime, this switches over the readiness check to use
`CodeCompletionsLegacy` like the application does.

This also adds a spec for the case when anthropic fails but vertex
succeeds. To do this, it refactors the specs for the readiness
endpoint a little bit so the mocks can be reused.