feat: only add readiness check in staging for now
We didn't notice the failure to start in staging, because I didn't
notice that the `runway.yml` file does not apply to both
environments. This changes the startup probe to only be tested in
staging first.

It also increases the timeout for the probe, as manual testing
revealed that the 2 calls can take about a second each.