chore: tweak startup probe in staging
We have quite a low quota for concurrent requests for Anthropic in
staging. So lowering the instances we start to 1, which should be
plenty for staging anyway. This reduces the number of instances we
start from `6 regions * 4 instances` = 24 to `6 regions * 1 instance =
6`.

Also slowing down the startup probe a bit.

It's not likely we'll need this in production as there we have much
higher concurrency limits.

For https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/issues/444