chore: use CLAUDE_INSTANT_1_2 for startup probe
The fastest, cheapest model Anthropic currently offers on the legacy
completions API. This should speed up querying and make it less likely
we hit concurrency limits.

This also adds a small valid prompt that is otherwise built in
GitLab-Rails. This seems to speed up the inference when trying
locally.

Related to #444