fix(deps): drop `always-copy` virtualenv option
This option was added in
https://gitlab.com/gitlab-org/modelops/applied-ml/code-suggestions/ai-assist/-/merge_requests/713
to ensure the Python interpreter exists even if Homebrew or `asdf`
removes the file. However, this option doesn't necessarily help
because it doesn't copy the Python shared library over. If a user
uninstalls the needed Python version then the virtualenv will be
broken in any case. In addition, `mise` doesn't work with this option
since the shared library is a relative path. See
https://github.com/mise-plugins/mise-poetry/issues/16 for more
details.